merge dev dans test

+++ core/src/main/java/fr/abes/item/core/components/FichierSauvegardeSuppCsv.java
@@ -13,10 +13,7 @@ import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
+import java.io.*;
 import java.nio.file.Path;
 import java.util.List;
 
@@ -38,10 +35,11 @@ public class FichierSauvegardeSuppCsv extends AbstractFichier implements Fichier
              PrintWriter out = new PrintWriter(bw)) {
             // création de la liste de référence pour trouver l'emplacement de chaque zone et sous-zone
             List<String> listDeReference = referenceService.constructHeaderCsv();
-            listDeReference.remove(0);
-            listDeReference.remove(0);
+            listDeReference.subList(0, 4).clear();
             // ajout de la ligne
-            out.println(typeDoc + ";" + ppn + ";" + gererZones(listDeReference, exemplaire));
+            String rcr = exemplaire.findZone("930", 0).findSubLabel("$b");
+            String epn = exemplaire.findZone("A99", 0).getValeur();
+            out.println(typeDoc + ";" + ppn + ";" + rcr + ";" + epn + ";" + gererZones(listDeReference, exemplaire));
         } catch (IOException ex) {
             throw new StorageException("Impossible d'écrire dans le fichier de sauvegarde csv");
         }
@@ -110,13 +108,25 @@ public class FichierSauvegardeSuppCsv extends AbstractFichier implements Fichier
     }
 
     public void writeHeader() {
-        try (FileWriter fw = new FileWriter(this.getPath().resolve(this.getFilename()).toString(), true);
-             BufferedWriter bw = new BufferedWriter(fw);
-             PrintWriter out = new PrintWriter(bw)) {
-            // ajout de la ligne
-            out.println(String.join(";", this.referenceService.constructHeaderCsv()));
+        if (!isHeaderExist()) {
+            try (FileWriter fw = new FileWriter(this.getPath().resolve(this.getFilename()).toString(), true);
+                 BufferedWriter bw = new BufferedWriter(fw);
+                 PrintWriter out = new PrintWriter(bw)) {
+                // ajout de la ligne
+                out.println(String.join(";", this.referenceService.constructHeaderCsv()));
+            } catch (IOException ex) {
+                throw new StorageException(Constant.ERR_FILE_WRITING + "de sauvegarde csv");
+            }
+        }
+    }
+
+    private boolean isHeaderExist() {
+        try (FileReader fr = new FileReader(this.getPath().resolve(this.getFilename()).toString());
+             BufferedReader br = new BufferedReader(fr)) {
+            String line = br.readLine();
+            return line != null && line.startsWith("TYPE (008);PPN;");
         } catch (IOException ex) {
-            throw new StorageException("Impossible d'écrire dans le fichier de sauvegarde csv");
+            throw new StorageException(Constant.ERR_FILE_READING + " de sauvegarde csv");
         }
     }
 }
